Create scenarios template
