+++ graphrag/query/cli.py
@@ -45,66 +45,9 @@ from .indexer_adapters import (
 
 from common.graph_db_client import GraphDBClient
 
-reporter = PrintProgressReporter("")
+from graphrag.query.scenario.runner import execute_scenario
 
 reporter = PrintProgressReporter("")
-
-def __get_embedding_description_store(
-    entities: list[Entity] = [],
-    vector_store_type: str = VectorStoreType.LanceDB,
-    config_args: dict | None = None,
-    context_id: str = "",
-):
-    """Get the embedding description store."""
-    if not config_args:
-        config_args = {}
-
-    collection_name = config_args.get(
-        "query_collection_name", "entity_description_embeddings"
-    )
-    config_args.update({"collection_name": f"{collection_name}_{context_id}" if context_id else collection_name})
-    vector_name = config_args.get(
-        "vector_search_column", "description_embedding"
-    )
-    config_args.update({"vector_name": vector_name})
-    config_args.update({"reports_name": f"reports_{context_id}" if context_id else "reports"})
-    config_args.update({"text_units_name": f"text_units_{context_id}"})
-
-    description_embedding_store = VectorStoreFactory.get_vector_store(
-        vector_store_type=vector_store_type, kwargs=config_args
-    )
-
-    description_embedding_store.connect(**config_args)
-
-    if vector_store_type == VectorStoreType.Kusto:
-        return description_embedding_store
-
-    elif config_args.get("overwrite", True):
-        # this step assumps the embeddings where originally stored in a file rather
-        # than a vector database
-
-        # dump embeddings from the entities list to the description_embedding_store
-        store_entity_semantic_embeddings(
-            entities=entities, vectorstore=description_embedding_store
-        )
-    else:
-        # load description embeddings to an in-memory lancedb vectorstore
-        # to connect to a remote db, specify url and port values.
-        description_embedding_store = LanceDBVectorStore(
-            collection_name=collection_name
-        )
-        description_embedding_store.connect(
-            db_uri=config_args.get("db_uri", "./lancedb")
-        )
-
-        # load data from an existing table
-        description_embedding_store.document_collection = (
-            description_embedding_store.db_connection.open_table(
-                description_embedding_store.collection_name
-            )
-        )
-
-    return description_embedding_store
 
 
 def run_global_search(
@@ -153,216 +96,6 @@ def run_global_search(
     reporter.success(f"Global Search Response: {result.response}")
     return result.response
 
-def path0(
-    config_dir: str | None,
-    data_dir: str | None,
-    root_dir: str | None,
-    community_level: int,
-    response_type: str,
-    context_id: str,
-    query: str,
-    optimized_search: bool = False,
-    use_kusto_community_reports: bool = False,
-    ):
-
-    """Run a local search with the given query."""
-    data_dir, root_dir, config = _configure_paths_and_settings(
-        data_dir, root_dir, config_dir
-    )
-
-    vector_store_args = (
-        config.embeddings.vector_store if config.embeddings.vector_store else {}
-    )
-
-    reporter.info(f"Vector Store Args: {vector_store_args}")
-    vector_store_type = vector_store_args.get("type", VectorStoreType.LanceDB)
-
-    entities=[]
-    text_units=[]
-    covariates=[]
-    reports=[]
-    final_relationships=[]
-
-    if(config.storage.type == StorageType.blob):
-        if(config.storage.container_name is not None):
-             output_storage_client: PipelineStorage = BlobPipelineStorage(connection_string=config.storage.connection_string,
-                                                                          container_name=config.storage.container_name,
-                                                                          storage_account_blob_url=config.storage.storage_account_blob_url)
-        else:
-            ValueError("Storage type is Blob but container name is invalid")
-    elif(config.storage.type == StorageType.file):
-        output_storage_client: PipelineStorage = FilePipelineStorage(config.root_dir)
-
-
-
-    ##### LEGACY #######################
-
-    if vector_store_type == VectorStoreType.LanceDB:
-        # for the POC purpose input artifacts blob, output artifacts blob and input query blob storage are going to same.
-        if(config.storage.type == StorageType.memory):
-            ValueError("Memory storage is not supported")
-        if(config.storage.type == StorageType.blob):
-            if(config.storage.container_name is not None):
-                input_storage_client: PipelineStorage = BlobPipelineStorage(connection_string=config.storage.connection_string,
-                                                                            container_name=config.storage.container_name,
-                                                                            storage_account_blob_url=config.storage.storage_account_blob_url)
-            else:
-                ValueError("Storage type is Blob but container name is invalid")
-        if(config.storage.type == StorageType.file):
-            input_storage_client: PipelineStorage = FilePipelineStorage(config.root_dir)
-
-
-        data_paths = []
-        data_paths = get_files_by_contextid(config, context_id)
-        final_nodes = pd.DataFrame()
-        final_community_reports = pd.DataFrame()
-        final_text_units = pd.DataFrame()
-        final_relationships = pd.DataFrame()
-        final_entities = pd.DataFrame()
-        final_covariates = pd.DataFrame()
-
-        for data_path in data_paths:
-            #check from the config for the ouptut storage type and then read the data from the storage.
-
-            #GraphDB: we may need to make change below to read nodes data from Graph DB
-            final_nodes = pd.concat([final_nodes, read_paraquet_file(input_storage_client, data_path + "/create_final_nodes.parquet")])
-            final_community_reports = pd.concat([final_community_reports,read_paraquet_file(input_storage_client, data_path + "/create_final_community_reports.parquet")]) # KustoDB: Final_entities, Final_Nodes, Final_report should be merged and inserted to kusto
-            final_text_units = pd.concat([final_text_units, read_paraquet_file(input_storage_client, data_path + "/create_final_text_units.parquet")]) # lance db search need it for embedding mapping. we have embeddings in entities we should use from there. KustoDB already must have sorted it.
-            final_relationships = pd.concat([final_relationships,read_paraquet_file(input_storage_client, data_path + "/create_final_relationships.parquet")])
-
-            if not optimized_search:
-                final_covariates = pd.concat([final_covariates, read_paraquet_file(input_storage_client, data_path + "/create_final_covariates.parquet")])
-
-            final_entities = pd.concat([final_entities, read_paraquet_file(input_storage_client, data_path + "/create_final_entities.parquet")])
-
-        ############# End of for loop
-
-        entities = read_indexer_entities(final_nodes, final_entities, community_level) # KustoDB: read Final nodes data and entities data and merge it.
-        reports=read_indexer_reports(
-            final_community_reports, final_nodes, community_level
-        )
-
-        final_relationships=read_indexer_relationships(final_relationships)
-
-        covariates = (
-            read_indexer_covariates(final_covariates)
-            if final_covariates.empty is False
-            else []
-        )
-        text_units=read_indexer_text_units(final_text_units)
-
-
-    ########################################################################################
-
-    if use_kusto_community_reports:
-        raise ValueError("Using community reports is not supported.")
-
-    description_embedding_store = __get_embedding_description_store(
-        entities=entities,
-        vector_store_type=vector_store_type,
-        config_args=vector_store_args,
-        context_id=context_id,
-    )
-
-    '''
-        *** If KUSTO is enabled, both entities and final_relationships must be empty.
-    '''
-    search_engine = get_local_search_engine(
-        config,
-        reports=reports,
-        text_units=text_units,
-        entities=entities,
-        relationships=final_relationships,
-        covariates={"claims": covariates},
-        description_embedding_store=description_embedding_store,
-        response_type=response_type,
-        context_id=context_id,
-        is_optimized_search=optimized_search,
-        use_kusto_community_reports=use_kusto_community_reports,
-        graphdb_config=config.graphdb,
-    )
-
-    if optimized_search:
-        result = search_engine.optimized_search(query=query)
-    else:
-        result = search_engine.search(query=query)
-    for key in  result.context_data.keys():
-        asyncio.run(output_storage_client.set("query/output/"+ key +".paraquet", result.context_data[key].to_parquet())) #it shows as error in editor but not an error.
-    reporter.success(f"Local Search Response: {result.response}")
-    return result.response
-
-def path1(
-    config_dir: str | None,
-    data_dir: str | None,
-    root_dir: str | None,
-    community_level: int,
-    response_type: str,
-    context_id: str,
-    query: str,
-    optimized_search: bool = False,
-    use_kusto_community_reports: bool = False,
-    ):
-    ValueError("Not implemented")
-
-def path2(
-    config_dir: str | None,
-    data_dir: str | None,
-    root_dir: str | None,
-    community_level: int,
-    response_type: str,
-    context_id: str,
-    query: str,
-    optimized_search: bool = False,
-    use_kusto_community_reports: bool = False,
-    ):
-    """Path 2
-    Find all the emails sent to trader by Tim Belden
-    a. Query -> LLM -> Entity Extracted -> 5 entities -> Set A [TimBelden1]
-    b. Query -> LLM -> Embeddings -> Y [x1..... Xn]
-    c. Run the query on Kusto for embedding Y [x1.....xn] for entitYid in [TimBelden1]
-    4. Get the text units and get the response"""
-    data_dir, root_dir, config = _configure_paths_and_settings(
-        data_dir, root_dir, config_dir
-    )
-
-    # Populate args with dict of arguments for the LLM
-    args = {}
-    args['api_key'] = config.llm.api_key
-    args['type'] = config.llm.type
-    args['model'] = config.llm.model
-    args['model_supports_json'] = config.llm.model_supports_json
-    args['api_base'] = config.llm.api_base
-    args['api_version'] = config.llm.api_version
-    args['deployment_name'] = config.llm.deployment_name
-    llmm = {}
-    llmm['llm'] = args
-
-
-    result = asyncio.run(run_gi(
-        docs=[Document(text=query, id='0')],
-        entity_types=config.entity_extraction.entity_types,
-        reporter = None,
-        pipeline_cache=None,
-        args=llmm,
-    ))
-
-    print(result.entities)
-    exit(0)
-
-def path3(
-    config_dir: str | None,
-    data_dir: str | None,
-    root_dir: str | None,
-    community_level: int,
-    response_type: str,
-    context_id: str,
-    query: str,
-    optimized_search: bool = False,
-    use_kusto_community_reports: bool = False,
-    ):
-    ValueError("Not implemented")
-
-
 def run_local_search(
     config_dir: str | None,
     data_dir: str | None,
@@ -374,14 +107,11 @@ def run_local_search(
     optimized_search: bool = False,
     use_kusto_community_reports: bool = False,
     paths: int = 0,):
+
     """Run a local search with the given query."""
-    if(paths==1):
-        return path1(config_dir, data_dir, root_dir, community_level, response_type, context_id, query, optimized_search, use_kusto_community_reports)
-    elif(paths==2):
-        return path2(config_dir, data_dir, root_dir, community_level, response_type, context_id, query, optimized_search, use_kusto_community_reports)
-    elif(paths==3):
-        return path3(config_dir, data_dir, root_dir, community_level, response_type, context_id, query, optimized_search, use_kusto_community_reports)
-    return path0(config_dir, data_dir, root_dir, community_level, response_type, context_id, query, optimized_search, use_kusto_community_reports)
+    execute_scenario(config_dir, data_dir, root_dir, community_level, response_type, context_id, 
+                     query, optimized_search, use_kusto_community_reports,paths)
+    return
 
 def blob_exists(container_client, blob_name):
     blob_client = container_client.get_blob_client(blob_name)
