+++ graphrag/query/scenario/runner.py
@@ -1,0 +1,388 @@
+import asyncio
+import os
+from pathlib import Path
+from typing import cast
+from io import BytesIO
+
+from datashaper import VerbCallbacks
+from graphrag.common.progress.rich import RichProgressReporter
+from graphrag.common.storage import PipelineStorage, BlobPipelineStorage, FilePipelineStorage
+from graphrag.common.utils.context_utils import get_files_by_contextid
+from graphrag.config.enums import StorageType
+from azure.core.exceptions import ResourceNotFoundError
+
+import pandas as pd
+
+from graphrag.config import (
+    create_graphrag_config,
+    GraphRagConfig,
+)
+from graphrag.common.progress import PrintProgressReporter
+from graphrag.index.verbs.entities.extraction.strategies.graph_intelligence.run_graph_intelligence import run_gi
+from graphrag.index.verbs.entities.extraction.strategies.typing import Document
+from graphrag.model.entity import Entity
+from graphrag.query.input.loaders.dfs import (
+    store_entity_semantic_embeddings,
+)
+from graphrag.vector_stores import VectorStoreFactory, VectorStoreType
+from graphrag.vector_stores.base import BaseVectorStore
+from graphrag.vector_stores.lancedb import LanceDBVectorStore
+from graphrag.vector_stores.kusto import KustoVectorStore
+from ..factories import get_global_search_engine, get_local_search_engine
+from ..indexer_adapters import (
+    read_indexer_covariates,
+    read_indexer_entities,
+    read_indexer_relationships,
+    read_indexer_reports,
+    kt_read_indexer_reports,
+    read_indexer_text_units,
+)
+
+reporter = PrintProgressReporter("")
+
+def __get_embedding_description_store(
+    entities: list[Entity] = [],
+    vector_store_type: str = VectorStoreType.LanceDB,
+    config_args: dict | None = None,
+    context_id: str = "",
+):
+    """Get the embedding description store."""
+    if not config_args:
+        config_args = {}
+
+    collection_name = config_args.get(
+        "query_collection_name", "entity_description_embeddings"
+    )
+    config_args.update({"collection_name": f"{collection_name}_{context_id}" if context_id else collection_name})
+    vector_name = config_args.get(
+        "vector_search_column", "description_embedding"
+    )
+    config_args.update({"vector_name": vector_name})
+    config_args.update({"reports_name": f"reports_{context_id}" if context_id else "reports"})
+    config_args.update({"text_units_name": f"text_units_{context_id}"})
+
+    description_embedding_store = VectorStoreFactory.get_vector_store(
+        vector_store_type=vector_store_type, kwargs=config_args
+    )
+
+    description_embedding_store.connect(**config_args)
+
+    if vector_store_type == VectorStoreType.Kusto:
+        return description_embedding_store
+
+    elif config_args.get("overwrite", True):
+        # this step assumps the embeddings where originally stored in a file rather
+        # than a vector database
+
+        # dump embeddings from the entities list to the description_embedding_store
+        store_entity_semantic_embeddings(
+            entities=entities, vectorstore=description_embedding_store
+        )
+    else:
+        raise ValueError("__get_embedding_description_store(): Unsupported config")
+
+    return description_embedding_store
+
+def _configure_paths_and_settings(
+    data_dir: str | None,
+    root_dir: str | None,
+    config_dir: str | None,
+) -> tuple[str, str | None, GraphRagConfig]:
+    if data_dir is None and root_dir is None:
+        msg = "Either data_dir or root_dir must be provided."
+        raise ValueError(msg)
+    if data_dir is None:
+        data_dir = _infer_data_dir(cast(str, root_dir))
+    config = _create_graphrag_config(root_dir, config_dir)
+    return data_dir, root_dir, config
+
+
+def _infer_data_dir(root: str) -> str:
+    output = Path(root) / "output"
+    # use the latest data-run folder
+    if output.exists():
+        folders = sorted(output.iterdir(), key=os.path.getmtime, reverse=True)
+        if len(folders) > 0:
+            folder = folders[0]
+            return str((folder / "artifacts").absolute())
+    msg = f"Could not infer data directory from root={root}"
+    raise ValueError(msg)
+
+
+def _create_graphrag_config(
+    root: str | None,
+    config_dir: str | None,
+) -> GraphRagConfig:
+    """Create a GraphRag configuration."""
+    return _read_config_parameters(root or "./", config_dir)
+
+
+def _read_config_parameters(root: str, config: str | None):
+    _root = Path(root)
+    settings_yaml = (
+        Path(config)
+        if config and Path(config).suffix in [".yaml", ".yml"]
+        else _root / "settings.yaml"
+    )
+    if not settings_yaml.exists():
+        settings_yaml = _root / "settings.yml"
+
+    if settings_yaml.exists():
+        reporter.info(f"Reading settings from {settings_yaml}")
+        with settings_yaml.open(
+            "rb",
+        ) as file:
+            import yaml
+
+            data = yaml.safe_load(file.read().decode(encoding="utf-8", errors="strict"))
+            return create_graphrag_config(data, root)
+
+    settings_json = (
+        Path(config)
+        if config and Path(config).suffix == ".json"
+        else _root / "settings.json"
+    )
+    if settings_json.exists():
+        reporter.info(f"Reading settings from {settings_json}")
+        with settings_json.open("rb") as file:
+            import json
+
+            data = json.loads(file.read().decode(encoding="utf-8", errors="strict"))
+            return create_graphrag_config(data, root)
+
+    reporter.info("Reading settings from environment variables")
+    return create_graphrag_config(root_dir=root)
+
+def path0(
+    config_dir: str | None,
+    data_dir: str | None,
+    root_dir: str | None,
+    community_level: int,
+    response_type: str,
+    context_id: str,
+    query: str,
+    optimized_search: bool = False,
+    use_kusto_community_reports: bool = False,
+    ):
+
+    """Run a local search with the given query."""
+    data_dir, root_dir, config = _configure_paths_and_settings(
+        data_dir, root_dir, config_dir
+    )
+
+    vector_store_args = (
+        config.embeddings.vector_store if config.embeddings.vector_store else {}
+    )
+
+    reporter.info(f"Vector Store Args: {vector_store_args}")
+    vector_store_type = vector_store_args.get("type", VectorStoreType.LanceDB)
+
+    entities=[]
+    text_units=[]
+    covariates=[]
+    reports=[]
+    final_relationships=[]
+
+    if(config.storage.type == StorageType.blob):
+        if(config.storage.container_name is not None):
+             output_storage_client: PipelineStorage = BlobPipelineStorage(connection_string=config.storage.connection_string,
+                                                                          container_name=config.storage.container_name,
+                                                                          storage_account_blob_url=config.storage.storage_account_blob_url)
+        else:
+            ValueError("Storage type is Blob but container name is invalid")
+    elif(config.storage.type == StorageType.file):
+        output_storage_client: PipelineStorage = FilePipelineStorage(config.root_dir)
+
+
+
+    ##### LEGACY #######################
+
+    if vector_store_type == VectorStoreType.LanceDB:
+        # for the POC purpose input artifacts blob, output artifacts blob and input query blob storage are going to same.
+        if(config.storage.type == StorageType.memory):
+            ValueError("Memory storage is not supported")
+        if(config.storage.type == StorageType.blob):
+            if(config.storage.container_name is not None):
+                input_storage_client: PipelineStorage = BlobPipelineStorage(connection_string=config.storage.connection_string,
+                                                                            container_name=config.storage.container_name,
+                                                                            storage_account_blob_url=config.storage.storage_account_blob_url)
+            else:
+                ValueError("Storage type is Blob but container name is invalid")
+        if(config.storage.type == StorageType.file):
+            input_storage_client: PipelineStorage = FilePipelineStorage(config.root_dir)
+
+
+        data_paths = []
+        data_paths = get_files_by_contextid(config, context_id)
+        final_nodes = pd.DataFrame()
+        final_community_reports = pd.DataFrame()
+        final_text_units = pd.DataFrame()
+        final_relationships = pd.DataFrame()
+        final_entities = pd.DataFrame()
+        final_covariates = pd.DataFrame()
+
+        for data_path in data_paths:
+            #check from the config for the ouptut storage type and then read the data from the storage.
+
+            #GraphDB: we may need to make change below to read nodes data from Graph DB
+            final_nodes = pd.concat([final_nodes, read_paraquet_file(input_storage_client, data_path + "/create_final_nodes.parquet")])
+            final_community_reports = pd.concat([final_community_reports,read_paraquet_file(input_storage_client, data_path + "/create_final_community_reports.parquet")]) # KustoDB: Final_entities, Final_Nodes, Final_report should be merged and inserted to kusto
+            final_text_units = pd.concat([final_text_units, read_paraquet_file(input_storage_client, data_path + "/create_final_text_units.parquet")]) # lance db search need it for embedding mapping. we have embeddings in entities we should use from there. KustoDB already must have sorted it.
+            final_relationships = pd.concat([final_relationships,read_paraquet_file(input_storage_client, data_path + "/create_final_relationships.parquet")])
+
+            if not optimized_search:
+                final_covariates = pd.concat([final_covariates, read_paraquet_file(input_storage_client, data_path + "/create_final_covariates.parquet")])
+
+            final_entities = pd.concat([final_entities, read_paraquet_file(input_storage_client, data_path + "/create_final_entities.parquet")])
+
+        ############# End of for loop
+
+        entities = read_indexer_entities(final_nodes, final_entities, community_level) # KustoDB: read Final nodes data and entities data and merge it.
+        reports=read_indexer_reports(
+            final_community_reports, final_nodes, community_level
+        )
+
+        final_relationships=read_indexer_relationships(final_relationships)
+
+        covariates = (
+            read_indexer_covariates(final_covariates)
+            if final_covariates.empty is False
+            else []
+        )
+        text_units=read_indexer_text_units(final_text_units)
+
+
+    ########################################################################################
+
+    if use_kusto_community_reports:
+        raise ValueError("Using community reports is not supported.")
+
+    description_embedding_store = __get_embedding_description_store(
+        entities=entities,
+        vector_store_type=vector_store_type,
+        config_args=vector_store_args,
+        context_id=context_id,
+    )
+
+    '''
+        *** If KUSTO is enabled, both entities and final_relationships must be empty.
+    '''
+    search_engine = get_local_search_engine(
+        config,
+        reports=reports,
+        text_units=text_units,
+        entities=entities,
+        relationships=final_relationships,
+        covariates={"claims": covariates},
+        description_embedding_store=description_embedding_store,
+        response_type=response_type,
+        context_id=context_id,
+        is_optimized_search=optimized_search,
+        use_kusto_community_reports=use_kusto_community_reports,
+        graphdb_config=config.graphdb,
+    )
+
+    if optimized_search:
+        result = search_engine.optimized_search(query=query)
+    else:
+        result = search_engine.search(query=query)
+    for key in  result.context_data.keys():
+        asyncio.run(output_storage_client.set("query/output/"+ key +".paraquet", result.context_data[key].to_parquet())) #it shows as error in editor but not an error.
+    reporter.success(f"Local Search Response: {result.response}")
+    return result.response
+
+def path1(
+    config_dir: str | None,
+    data_dir: str | None,
+    root_dir: str | None,
+    community_level: int,
+    response_type: str,
+    context_id: str,
+    query: str,
+    optimized_search: bool = False,
+    use_kusto_community_reports: bool = False,
+    ):
+    ValueError("Not implemented")
+
+def path2(
+    config_dir: str | None,
+    data_dir: str | None,
+    root_dir: str | None,
+    community_level: int,
+    response_type: str,
+    context_id: str,
+    query: str,
+    optimized_search: bool = False,
+    use_kusto_community_reports: bool = False,
+    ):
+    """Path 2
+    Find all the emails sent to trader by Tim Belden
+    a. Query -> LLM -> Entity Extracted -> 5 entities -> Set A [TimBelden1]
+    b. Query -> LLM -> Embeddings -> Y [x1..... Xn]
+    c. Run the query on Kusto for embedding Y [x1.....xn] for entitYid in [TimBelden1]
+    4. Get the text units and get the response"""
+    data_dir, root_dir, config = _configure_paths_and_settings(
+        data_dir, root_dir, config_dir
+    )
+
+    # Populate args with dict of arguments for the LLM
+    args = {}
+    args['api_key'] = config.llm.api_key
+    args['type'] = config.llm.type
+    args['model'] = config.llm.model
+    args['model_supports_json'] = config.llm.model_supports_json
+    args['api_base'] = config.llm.api_base
+    args['api_version'] = config.llm.api_version
+    args['deployment_name'] = config.llm.deployment_name
+    llmm = {}
+    llmm['llm'] = args
+
+
+    result = asyncio.run(run_gi(
+        docs=[Document(text=query, id='0')],
+        entity_types=config.entity_extraction.entity_types,
+        reporter = None,
+        pipeline_cache=None,
+        args=llmm,
+    ))
+
+    print(result.entities)
+    exit(0)
+
+def path3(
+    config_dir: str | None,
+    data_dir: str | None,
+    root_dir: str | None,
+    community_level: int,
+    response_type: str,
+    context_id: str,
+    query: str,
+    optimized_search: bool = False,
+    use_kusto_community_reports: bool = False,
+    ):
+    ValueError("Not implemented")
+
+def execute_scenario(
+    config_dir: str | None,
+    data_dir: str | None,
+    root_dir: str | None,
+    community_level: int,
+    response_type: str,
+    context_id: str,
+    query: str,
+    optimized_search: bool = False,
+    use_kusto_community_reports: bool = False,
+    paths: int = 0,):
+
+    """Run a local search with the given query."""
+    if(paths==1):
+        return path1(config_dir, data_dir, root_dir, community_level, response_type, context_id, query, optimized_search, use_kusto_community_reports)
+    elif(paths==2):
+        return path2(config_dir, data_dir, root_dir, community_level, response_type, context_id, query, optimized_search, use_kusto_community_reports)
+    elif(paths==3):
+        return path3(config_dir, data_dir, root_dir, community_level, response_type, context_id, query, optimized_search, use_kusto_community_reports)
+    return path0(config_dir, data_dir, root_dir, community_level, response_type, context_id, query, optimized_search, use_kusto_community_reports)
+
+
+def test():
+    print("ok")
